리뷰이벤트리스너, 퍼블리셔 분리
